Fix issue with missing second and further chunks for a source when there are no chunks from other sources in parallel
1. Fix `ChunksMerger._transform` method to always push the chunk itself
   to the output side of the stream (including when current chunk's
   source is the same as the previous chunk's source)
2. Extract `ChunkStream` and `ChunksMerger` classes into separate files
