+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "merge-stream-in-chunks",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "description": "Merge several output streams minimizing data shuffling between the streams producing output in parallel",
   "main": "index.js",
   "scripts": {
